Fix integer overflow bug

+++ src/types/shift_break.cr
@@ -9,7 +9,7 @@ module Tanda::CLI
 
       getter id : Int32
       getter shift_id : Int32
-      getter length : UInt8 # length in minutes
+      getter length : UInt16 # length in minutes
 
       @[JSON::Field(key: "start", converter: Tanda::CLI::Types::Converters::Time::FromMaybeUnix)]
       getter start_time : Time?
